requirements.txt
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
urllib3>=2.0.0
python-whois>=0.8.0
dnspython>=2.4.0
requests>=2.31.0
certifi>=2023.7.22
